+++ Ejercicio 6/v5 dist. varios brokers con anillo y lider/broker/Broker.cpp
@@ -48,8 +48,8 @@ void crear_ipcs(int master){
 	sem_anillo.iniSem(0);
 	
 	key = ftok(ipcFileName.c_str(), SHM_VERSION);
-    int shmversion = shmget(key, sizeof(int) , 0660 | IPC_CREAT);
-    int* version_id = (int*)shmat(shmversion, NULL, 0);   
+    int shmversion = shmget(key, sizeof(uint64_t) , 0660 | IPC_CREAT);
+    uint64_t* version_id = (uint64_t*)shmat(shmversion, NULL, 0);   
     *version_id = ID_BROKER;
     shmdt((void*)version_id);
